Update APIs including authorization

+++ src/app/api/Admin/category/delete-category/route.ts
@@ -8,26 +8,30 @@ export async function DELETE(req: Request) {
     await connectDB();
     const isAuthenticated = await AuthCheck(req);
 
+    if (!isAuthenticated) {
+      return NextResponse.json({ success: false, message: "You are not authorized." });
+    }
+
     if (isAuthenticated === 'admin') {
       const { searchParams } = new URL(req.url);
       const id = searchParams.get('id');
 
       if (!id) {
-        return NextResponse.json({ success: true, message: "Category ID is Required" });
+        return NextResponse.json({ success: false, message: "Category ID is required." });
       }
 
       const deleteData = await Category.findByIdAndDelete(id);
 
       if (deleteData) {
-        return NextResponse.json({ success: true, message: "Category Deleted successfully!" });
+        return NextResponse.json({ success: true, message: "Category deleted successfully!" });
       } else {
-        return NextResponse.json({ success: false, message: "Failed to Delete the category. Please try again!" });
+        return NextResponse.json({ success: false, message: "Failed to delete the category. Please try again!" });
       }
     } else {
-      return NextResponse.json({ success: false, message: "You are not authorized." });
+      return NextResponse.json({ success: false, message: "No permission to delete the category." });
     }
   } catch (error) {
     console.log('Error in deleting a category:', error);
-    return NextResponse.json({ success: false, message: 'Something went wrong. Please try again!' });
+    return NextResponse.json({ status: 500, success: false, message: 'Something went wrong. Please try again!' });
   }
 }
